Fix request error handler for Node.JS (#9)
We currently initialize the error handler for the request in the
`NodeHttpClientSession` way too late - when we get first bytes of the
response. So, in case there's an error before we get any response (e.g.
because of a timeout) - the exception will be bubbled up uncaught to the
very top of the application stack.

To avoid that, let's make sure we initialize the error handler right
after creating the request.